Capsule-capsule collision detection as a Warp kernel
A prototype of capsule capsule collision detection
primitive in Warp.

This also contains a very rough plumbing to be able to call
from Jax to Warp, but many things are missing there;
for instance, there is no grad, no support for multiple
GPUs, there is a very simple minded and hacky API
with confusing contract, etc.

Needs `pip install warp-lang`